+++ ruby/CHANGELOG.md
@@ -5,8 +5,9 @@ This is the changelog for the `essence` gem. For now, most of the updates will b
 ### Unreleased
 
 - feature: Adding `Checkbox` component
-- feature: Adding `Switch` component
+- feature: Adding `Label` component
 - feature: Adding `Separator` component
+- feature: Adding `Switch` component
 
 ### 0.3.1